Retrieve a directory

Retrieve the path to a registered directory (or its subdirectories), possibly creating a local copy of the directory's contents if the caller is not on the same filesystem.

retrieveDirectory(
  path,
  url,
  cache = NULL,
  forceRemote = FALSE,
  overwrite = FALSE,
  concurrent = 1
)

Arguments

path

String containing the absolute path to a directory.

url

String containing the URL to the SewerRat REST API.

cache

String containing a path to a cache directory. If NULL, an appropriate location is automatically chosen. Only used for remote access.

forceRemote

Logical scalar indicating whether to force remote access. This will download all files in the path via the REST API and cache them locally, even if path is on the same filesystem as the caller.

overwrite

Logical scalar indicating whether to overwrite the existing cache. Only used for remote access.

concurrent

Integer specifying the number of concurrent downloads. Only used for remote access.

Value

Path to the subdirectory on the caller's filesystem. This is either a path to the registered (sub)directory if it is accessible, or a path to a local cache of the directory's contents otherwise.

Author

Aaron Lun

Examples

info <- startSewerRat()

# Mocking up a directory of stuff to query.
mydir <- tempfile()
dir.create(mydir)
write(file=file.path(mydir, "foo"), '{ "first": "Aaron", "last": "Lun" }')
dir.create(file.path(mydir, "diet"))
write(file=file.path(mydir, "diet", "bar"), 
   '{ "meal": "lunch", "ingredients": "water" }')

# Registering it:
register(mydir, "metadata.json", url=info$url)

# Fetching it, either directly or via the API.
retrieveDirectory(mydir, url=info$url)
#> [1] "/tmp/Rtmphnlz7W/file21a55226990"
retrieveDirectory(paste0(mydir, "/diet"), url=info$url, forceRemote=TRUE)
#> [1] "/github/home/.local/share/R/SewerRat/http%3A%2F%2F0.0.0.0%3A5114/LOCAL//tmp/Rtmphnlz7W/file21a55226990/diet"